requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-spectacular==0.22.0
idna==3.3
importlib-resources==5.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
